+++ src/bl64-rbac-env.bash
@@ -1,7 +1,7 @@
 #######################################
 # BashLib64 / Module / Globals / Manage role based access service
 #
-# Version: 1.5.0
+# Version: 1.6.0
 #######################################
 
 export BL64_RBAC_MODULE="$BL64_VAR_OFF"
@@ -11,6 +11,10 @@ export BL64_RBAC_CMD_VISUDO=''
 export BL64_RBAC_FILE_SUDOERS=''
 
 export BL64_RBAC_ALIAS_SUDO_ENV=''
+
+export BL64_RBAC_SET_SUDO_CHECK=''
+export BL64_RBAC_SET_SUDO_FILE=''
+export BL64_RBAC_SET_SUDO_QUIET=''
 
 export _BL64_RBAC_TXT_INVALID_SUDOERS='the sudoers file is corrupt or invalid'
 export _BL64_RBAC_TXT_ADD_ROOT='add password-less root privilege to user'