+++ src/bl64-rbac.bash
@@ -1,7 +1,7 @@
 #######################################
 # BashLib64 / Module / Functions / Manage role based access service
 #
-# Version: 1.10.1
+# Version: 1.11.0
 #######################################
 
 #######################################
@@ -31,6 +31,15 @@ function bl64_rbac_add_root() {
 
   bl64_msg_show_lib_task "$_BL64_RBAC_TXT_ADD_ROOT ($user)"
   umask 0266
+
+  if [[ -s "$BL64_RBAC_FILE_SUDOERS" ]]; then
+    bl64_dbg_lib_show_info "backup original sudoers (${BL64_RBAC_FILE_SUDOERS} -> ${old_sudoers})"
+    bl64_fs_cp_file "${BL64_RBAC_FILE_SUDOERS}" "$old_sudoers"
+    status=$?
+    ((status != 0)) && bl64_msg_show_error "unable to backup sudoers file (${BL64_RBAC_FILE_SUDOERS})" && return $status
+  fi
+
+  bl64_dbg_lib_show_info "create new sudoers (${new_sudoers})"
   # shellcheck disable=SC2016
   bl64_txt_run_awk \
     -v ControlUsr="$user" \
@@ -45,19 +54,14 @@ function bl64_rbac_add_root() {
       }
     ' \
     "$BL64_RBAC_FILE_SUDOERS" >"$new_sudoers"
+  status=$?
+  ((status != 0)) && bl64_msg_show_error "unable to create new sudoers file (${new_sudoers})" && return $status
 
-  if [[ -s "$new_sudoers" ]]; then
-    bl64_fs_cp_file "${BL64_RBAC_FILE_SUDOERS}" "$old_sudoers"
-  fi
-  if [[ -s "$new_sudoers" && -s "$old_sudoers" ]]; then
-    "$BL64_OS_CMD_CAT" "${BL64_RBAC_FILE_SUDOERS}.bl64_new" >"${BL64_RBAC_FILE_SUDOERS}" &&
-      bl64_rbac_check_sudoers "$BL64_RBAC_FILE_SUDOERS"
-    status=$?
-  else
-    status=$BL64_LIB_ERROR_TASK_FAILED
-  fi
+  bl64_dbg_lib_show_info "replace original sudoers with new version (${new_sudoers} ->${BL64_RBAC_FILE_SUDOERS})"
+  "$BL64_OS_CMD_CAT" "$new_sudoers" >"${BL64_RBAC_FILE_SUDOERS}" &&
+    bl64_rbac_check_sudoers "$BL64_RBAC_FILE_SUDOERS"
+  status=$?
 
-  # shellcheck disable=SC2086
   return $status
 }
 
@@ -77,7 +81,7 @@ function bl64_rbac_check_sudoers() {
   bl64_dbg_lib_show_function "$@"
   local sudoers="$1"
   local -i status=0
-  local debug='--quiet'
+  local debug="$BL64_RBAC_SET_SUDO_QUIET"
 
   bl64_check_parameter 'sudoers' &&
     bl64_check_command "$BL64_RBAC_CMD_VISUDO" ||
@@ -88,8 +92,8 @@ function bl64_rbac_check_sudoers() {
   # shellcheck disable=SC2086
   "$BL64_RBAC_CMD_VISUDO" \
     $debug \
-    --check \
-    --file="$sudoers"
+    $BL64_RBAC_SET_SUDO_CHECK \
+    ${BL64_RBAC_SET_SUDO_FILE}="$sudoers"
   status=$?
 
   if ((status != 0)); then